Hackaton11/jcondezo/requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
httpie==3.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
psycopg2==2.9.6
Pygments==2.15.1
PySocks==1.7.1
python-slugify==8.0.1
requests==2.29.0
requests-toolbelt==1.0.0
rich==13.3.5
SQLAlchemy==2.0.12
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.3
WTForms==3.0.1